Submit

+++ submission/src/LLL.h
@@ -1,6 +1,6 @@
 #ifndef LLL_H
 #define LLL_H
 
-void LLL(double **B, double **u, double **mu, int delta, int N);
+void LLL(double **B, double **u, double **mu, double delta, int N);
 
 #endif
